Implement actions and reducers for view article page
